parsers/dash: add parseRepresentations.ts
