+++ src/parsers/manifest/dash/parseAdaptationSets.ts
@@ -15,78 +15,26 @@
  */
 
 import arrayFind from "array-find";
-import log from "../../../log";
-import IRepresentationIndex from "../../../manifest/representation_index";
 import { normalize as normalizeLang } from "../../../utils/languages";
 import { resolveURL } from "../../../utils/url";
 import {
-  IContentProtection,
   IParsedAdaptation,
   IParsedAdaptations,
-  IParsedRepresentation,
 } from "../types";
 import inferAdaptationType from "../utils/infer_adaptation_type";
 import {
   isHardOfHearing,
   isVisuallyImpaired,
 } from "./helpers";
-import BaseRepresentationIndex from "./indexes/base";
-import ListRepresentationIndex from "./indexes/list";
-import TemplateRepresentationIndex from "./indexes/template";
-import TimelineRepresentationIndex from "./indexes/timeline";
 import {
   IAdaptationSetIntermediateRepresentation
 } from "./node_parsers/AdaptationSet";
+import parseRepresentations from "./parseRepresentations";
 
 export interface IPeriodInfos {
   isDynamic : boolean;
   start : number;
   baseURL? : string;
-}
-
-// base context given to the various indexes
-interface IIndexContext {
-  periodStart : number; // Start of the period concerned by this
-                        // RepresentationIndex, in seconds
-  representationBaseURL : string; // Base URL for the Representation concerned
-  representationId? : string; // ID of the Representation concerned
-  representationBitrate? : number; // Bitrate of the Representation concerned
-  isDynamic : boolean;
-}
-
-/**
- * Find and parse RepresentationIndex located in an AdaptationSet node.
- * Returns a generic parsed SegmentTemplate with a single element if not found.
- * @param {Object} adaptation
- * @param {Object} context
- */
-function findAdaptationIndex(
-  adaptation : IAdaptationSetIntermediateRepresentation,
-  context: IIndexContext
-): IRepresentationIndex {
-  const adaptationChildren = adaptation.children;
-  let adaptationIndex : IRepresentationIndex;
-  if (adaptationChildren.segmentBase != null) {
-    const { segmentBase } = adaptationChildren;
-    adaptationIndex = new BaseRepresentationIndex(segmentBase, context);
-  } else if (adaptationChildren.segmentList != null) {
-    const { segmentList } = adaptationChildren;
-    adaptationIndex = new ListRepresentationIndex(segmentList, context);
-  } else if (adaptationChildren.segmentTemplate != null) {
-    const { segmentTemplate } = adaptationChildren;
-    adaptationIndex = segmentTemplate.indexType === "timeline" ?
-      new TimelineRepresentationIndex(segmentTemplate, context) :
-      new TemplateRepresentationIndex(segmentTemplate, context);
-  } else {
-    adaptationIndex = new TemplateRepresentationIndex({
-      duration: Number.MAX_VALUE,
-      timescale: 1,
-      startNumber: 0,
-      initialization: { media: "" },
-      media: "",
-    }, context);
-  }
-  return adaptationIndex;
 }
 
 /**
@@ -101,180 +49,18 @@ export default function parseAdaptationSets(
 ): IParsedAdaptations {
   return adaptationsIR
     .reduce<{
-      videoMainAdaptation : IParsedAdaptation|null;
       adaptations : IParsedAdaptations;
+      videoMainAdaptation : IParsedAdaptation|null;
     }>((acc, adaptation) => {
       const adaptationChildren = adaptation.children;
       const parsedAdaptations = acc.adaptations;
-      const adaptationBaseURL =
-        resolveURL(periodInfos.baseURL, adaptationChildren.baseURL);
 
-      // 4-1. Construct Representations
-      const representations = adaptation.children
-        .representations.map((representation) => {
-          const baseURL = representation.children.baseURL;
-          const representationBaseURL = resolveURL(adaptationBaseURL, baseURL);
-
-          // 4-2-1. Find Index
-          const context = {
-            periodStart: periodInfos.start,
-            isDynamic: periodInfos.isDynamic,
-            representationBaseURL,
-            representationId: representation.attributes.id,
-            representationBitrate: representation.attributes.bitrate,
-          };
-          let representationIndex : IRepresentationIndex;
-          if (representation.children.segmentBase != null) {
-            const { segmentBase } = representation.children;
-            representationIndex = new BaseRepresentationIndex(segmentBase, context);
-          } else if (representation.children.segmentList != null) {
-            const { segmentList } = representation.children;
-            representationIndex = new ListRepresentationIndex(segmentList, context);
-          } else if (representation.children.segmentTemplate != null) {
-            const { segmentTemplate } = representation.children;
-            representationIndex = segmentTemplate.indexType === "timeline" ?
-              new TimelineRepresentationIndex(segmentTemplate, context) :
-              new TemplateRepresentationIndex(segmentTemplate, context);
-          } else {
-            representationIndex = findAdaptationIndex(adaptation, context);
-          }
-
-          // 4-2-2. Find bitrate
-          let representationBitrate : number;
-          if (representation.attributes.bitrate == null) {
-            log.warn("DASH: No usable bitrate found in the Representation.");
-            representationBitrate = 0;
-          } else {
-            representationBitrate = representation.attributes.bitrate;
-          }
-
-          // 4-2-3. Set ID
-          const representationID = representation.attributes.id != null ?
-            representation.attributes.id :
-            (
-              representation.attributes.bitrate +
-              (
-                representation.attributes.height != null ?
-                ("-" + representation.attributes.height) : ""
-              ) +
-              (
-                representation.attributes.width != null ?
-                ("-" + representation.attributes.width) : ""
-              ) +
-              (
-                representation.attributes.mimeType != null ?
-                ("-" + representation.attributes.mimeType) : ""
-              ) +
-              (
-                representation.attributes.codecs != null ?
-                ("-" + representation.attributes.codecs) : ""
-              )
-            );
-          // 4-2-4. Construct Representation Base
-          const parsedRepresentation : IParsedRepresentation = {
-            bitrate: representationBitrate,
-            index: representationIndex,
-            id: representationID,
-          };
-          // 4-2-5. Add optional attributes
-          let codecs : string|undefined;
-          if (representation.attributes.codecs != null) {
-            codecs = representation.attributes.codecs;
-          } else if (adaptation.attributes.codecs != null) {
-            codecs = adaptation.attributes.codecs;
-          }
-          if (codecs != null) {
-            codecs = codecs === "mp4a.40.02" ? "mp4a.40.2" : codecs;
-            parsedRepresentation.codecs = codecs;
-          }
-          if (representation.attributes.audioSamplingRate != null) {
-            parsedRepresentation.audioSamplingRate =
-              representation.attributes.audioSamplingRate;
-          } else if (adaptation.attributes.audioSamplingRate != null) {
-            parsedRepresentation.audioSamplingRate =
-              adaptation.attributes.audioSamplingRate;
-          }
-          if (representation.attributes.codingDependency != null) {
-            parsedRepresentation.codingDependency =
-              representation.attributes.codingDependency;
-          } else if (adaptation.attributes.codingDependency != null) {
-            parsedRepresentation.codingDependency =
-              adaptation.attributes.codingDependency;
-          }
-          if (representation.attributes.frameRate != null) {
-            parsedRepresentation.frameRate =
-              representation.attributes.frameRate;
-          } else if (adaptation.attributes.frameRate != null) {
-            parsedRepresentation.frameRate =
-              adaptation.attributes.frameRate;
-          }
-          if (representation.attributes.height != null) {
-            parsedRepresentation.height =
-              representation.attributes.height;
-          } else if (adaptation.attributes.height != null) {
-            parsedRepresentation.height =
-              adaptation.attributes.height;
-          }
-          if (representation.attributes.maxPlayoutRate != null) {
-            parsedRepresentation.maxPlayoutRate =
-              representation.attributes.maxPlayoutRate;
-          } else if (adaptation.attributes.maxPlayoutRate != null) {
-            parsedRepresentation.maxPlayoutRate =
-              adaptation.attributes.maxPlayoutRate;
-          }
-          if (representation.attributes.maximumSAPPeriod != null) {
-            parsedRepresentation.maximumSAPPeriod =
-              representation.attributes.maximumSAPPeriod;
-          } else if (adaptation.attributes.maximumSAPPeriod != null) {
-            parsedRepresentation.maximumSAPPeriod =
-              adaptation.attributes.maximumSAPPeriod;
-          }
-          if (representation.attributes.mimeType != null) {
-            parsedRepresentation.mimeType =
-              representation.attributes.mimeType;
-          } else if (adaptation.attributes.mimeType != null) {
-            parsedRepresentation.mimeType =
-              adaptation.attributes.mimeType;
-          }
-          if (representation.attributes.profiles != null) {
-            parsedRepresentation.profiles =
-              representation.attributes.profiles;
-          } else if (adaptation.attributes.profiles != null) {
-            parsedRepresentation.profiles =
-              adaptation.attributes.profiles;
-          }
-          if (representation.attributes.qualityRanking != null) {
-            parsedRepresentation.qualityRanking =
-              representation.attributes.qualityRanking;
-          }
-          if (representation.attributes.segmentProfiles != null) {
-            parsedRepresentation.segmentProfiles =
-              representation.attributes.segmentProfiles;
-          } else if (adaptation.attributes.segmentProfiles != null) {
-            parsedRepresentation.segmentProfiles =
-              adaptation.attributes.segmentProfiles;
-          }
-          if (representation.attributes.width != null) {
-            parsedRepresentation.width =
-              representation.attributes.width;
-          } else if (adaptation.attributes.width != null) {
-            parsedRepresentation.width =
-              adaptation.attributes.width;
-          }
-          if (adaptation.children.contentProtections) {
-            const contentProtections : IContentProtection[] = [];
-            for (let k = 0; k < adaptation.children.contentProtections.length; k++) {
-              const protection = adaptation.children.contentProtections[k];
-              if (protection.keyId != null) {
-                contentProtections.push({ keyId: protection.keyId });
-              }
-            }
-            if (contentProtections.length) {
-              parsedRepresentation.contentProtections = contentProtections;
-            }
-          }
-          return parsedRepresentation;
-        });
+      const representationsIR = adaptation.children.representations;
+      const representations = parseRepresentations(representationsIR, adaptation, {
+        isDynamic: periodInfos.isDynamic,
+        start: periodInfos.start,
+        baseURL: resolveURL(periodInfos.baseURL, adaptationChildren.baseURL),
+      });
       const adaptationMimeType = adaptation.attributes.mimeType;
       const adaptationCodecs = adaptation.attributes.codecs;
       const representationMimeTypes = representations
